# docs/接口文档/接口json/catalog-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "游客端可售内容查询接口",
    "description": "游客查询景点、附近景点、景点资料以及指定日期的可展示场次和票种。",
    "version": "1.1.0"
  },
  "tags": [
    {
      "name": "Catalog",
      "description": "游客端公开可售内容查询"
    }
  ],
  "security": [],
  "paths": {
    "/public/venues": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "分页查询景点",
        "description": "查询全部启用景点；没有当前可预约场次时 minimumPrice 为 null，可按景点名称或地址搜索。",
        "operationId": "pageVenues",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Size"
          },
          {
            "$ref": "#/components/parameters/Keyword"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenuePageApiResponse"
                },
                "example": {
                  "code": "SUCCESS",
                  "message": "操作成功",
                  "data": {
                    "items": [
                      {
                        "id": 10,
                        "name": "海湾科技馆",
                        "address": "广州市番禺区海湾大道 88 号",
                        "description": "面向公众开放的综合性科技文化场馆。",
                        "coverUrl": "https://example.com/venue-cover.jpg",
                        "minimumPrice": 0.00
                      }
                    ],
                    "total": 1,
                    "page": 1,
                    "size": 20
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          }
        }
      }
    },
    "/public/venues/nearby": {
      "get": {
        "tags": ["Catalog"],
        "summary": "根据当前坐标查询附近景点",
        "description": "以游客当前坐标为中心查询指定半径内的启用景点，并按距离从近到远返回。",
        "operationId": "listNearbyVenues",
        "parameters": [
          {
            "name": "longitude",
            "in": "query",
            "required": true,
            "description": "查询中心经度",
            "schema": {"type": "number", "format": "double", "minimum": -180, "maximum": 180}
          },
          {
            "name": "latitude",
            "in": "query",
            "required": true,
            "description": "查询中心纬度",
            "schema": {"type": "number", "format": "double", "minimum": -90, "maximum": 90}
          },
          {
            "$ref": "#/components/parameters/RadiusKm"
          },
          {
            "$ref": "#/components/parameters/NearbyLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功；没有附近景点时 data 为空数组",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/NearbyVenueListApiResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequest"}
        }
      }
    },
    "/public/venues/nearby/by-name": {
      "get": {
        "tags": ["Catalog"],
        "summary": "根据地点名称查询附近景点",
        "description": "先通过高德地理编码服务将地点名称转换为坐标，再查询附近景点。",
        "operationId": "listNearbyVenuesByName",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "description": "地点名称或地址",
            "schema": {"type": "string", "maxLength": 100, "example": "广州塔"}
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "description": "可选城市，用于减少同名地点歧义",
            "schema": {"type": "string", "maxLength": 50, "example": "广州"}
          },
          {
            "$ref": "#/components/parameters/RadiusKm"
          },
          {
            "$ref": "#/components/parameters/NearbyLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功；没有附近景点时 data 为空数组",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/NearbyVenueListApiResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/InvalidRequest"}
        }
      }
    },
    "/public/venues/{venueId}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "获取景点详情",
        "description": "获取启用景点的基础展示资料。景点暂时售罄时仍可返回详情。",
        "operationId": "getPublicVenue",
        "parameters": [
          {
            "$ref": "#/components/parameters/VenueId"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenueDetailApiResponse"
                },
                "example": {
                  "code": "SUCCESS",
                  "message": "操作成功",
                  "data": {
                    "id": 10,
                    "name": "海湾科技馆",
                    "address": "广州市番禺区海湾大道 88 号",
                    "description": "面向公众开放的综合性科技文化场馆。",
                    "coverUrl": "https://example.com/venue-cover.jpg"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/VenueNotFound"
          }
        }
      }
    },
    "/public/venues/{venueId}/sessions": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "按日期查询可展示场次",
        "description": "查询启用景点在指定日期已经发布且预约尚未结束的场次。未开售场次只返回静态资料，开售后额外返回动态库存。",
        "operationId": "listSellableSessions",
        "parameters": [
          {
            "$ref": "#/components/parameters/VenueId"
          },
          {
            "$ref": "#/components/parameters/VisitDate"
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功；没有可展示场次时 data 为空数组",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListApiResponse"
                },
                "example": {
                  "code": "SUCCESS",
                  "message": "操作成功",
                  "data": [
                    {
                      "id": 21,
                      "visitDate": "2026-08-03",
                      "startTime": "09:00:00",
                      "endTime": "11:00:00",
                      "bookingStartAt": "2026-08-01T10:00:00",
                      "bookingEndAt": "2026-08-03T08:30:00",
                      "saleState": "ON_SALE",
                      "remainingCapacity": 24,
                      "ticketTypes": [
                        {
                          "sessionTicketTypeId": 301,
                          "ticketTypeName": "成人票",
                          "description": "普通参观票",
                          "audienceRule": "普通成年游客",
                          "salePrice": 60.00,
                          "remainingQuantity": 20
                        },
                        {
                          "sessionTicketTypeId": 302,
                          "ticketTypeName": "儿童票",
                          "description": "儿童优惠票",
                          "audienceRule": "1.2 米以下儿童",
                          "salePrice": 0.00,
                          "remainingQuantity": 4
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "404": {
            "$ref": "#/components/responses/VenueNotFound"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "页码，从 1 开始",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Size": {
        "name": "size",
        "in": "query",
        "required": false,
        "description": "每页数量",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "Keyword": {
        "name": "keyword",
        "in": "query",
        "required": false,
        "description": "景点名称或地址关键词",
        "schema": {
          "type": "string",
          "maxLength": 100
        }
      },
      "RadiusKm": {
        "name": "radiusKm",
        "in": "query",
        "required": false,
        "description": "搜索半径，单位为公里",
        "schema": {
          "type": "number",
          "format": "double",
          "minimum": 0.1,
          "maximum": 50,
          "default": 5
        }
      },
      "NearbyLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "最大返回数量",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "VenueId": {
        "name": "venueId",
        "in": "path",
        "required": true,
        "description": "景点 ID",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
      "VisitDate": {
        "name": "visitDate",
        "in": "query",
        "required": true,
        "description": "参观日期，不能早于当前日期",
        "schema": {
          "type": "string",
          "format": "date",
          "example": "2026-08-03"
        }
      }
    },
    "schemas": {
      "VenueCard": {
        "type": "object",
        "required": [
          "id",
          "name",
          "address",
          "minimumPrice"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "景点 ID"
          },
          "name": {
            "type": "string",
            "description": "景点名称"
          },
          "address": {
            "type": "string",
            "description": "景点地址"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "景点简介"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true,
            "description": "封面图片访问地址"
          },
          "minimumPrice": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "nullable": true,
            "description": "当前可售场次票种的最低实际售价；暂不可预约时为 null"
          }
        }
      },
      "VenueDetail": {
        "type": "object",
        "required": [
          "id",
          "name",
          "address"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "景点 ID"
          },
          "name": {
            "type": "string",
            "description": "景点名称"
          },
          "address": {
            "type": "string",
            "description": "景点地址"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "景点简介"
          },
          "coverUrl": {
            "type": "string",
            "nullable": true,
            "description": "封面图片访问地址"
          }
        }
      },
      "NearbyVenue": {
        "type": "object",
        "required": ["id", "name", "address", "distanceKm"],
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "name": {"type": "string"},
          "address": {"type": "string"},
          "description": {"type": "string", "nullable": true},
          "coverUrl": {"type": "string", "nullable": true},
          "distanceKm": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "description": "距查询中心的公里数"
          }
        }
      },
      "SellableSession": {
        "type": "object",
        "required": [
          "id",
          "visitDate",
          "startTime",
          "endTime",
          "bookingStartAt",
          "bookingEndAt",
          "saleState",
          "ticketTypes"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "场次 ID"
          },
          "visitDate": {
            "type": "string",
            "format": "date",
            "description": "参观日期"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "场次开始时间"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "场次结束时间"
          },
          "bookingStartAt": {
            "type": "string",
            "format": "date-time",
            "description": "预约开始时间"
          },
          "bookingEndAt": {
            "type": "string",
            "format": "date-time",
            "description": "预约截止时间"
          },
          "saleState": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "ON_SALE",
              "SOLD_OUT"
            ],
            "description": "游客端场次销售展示状态"
          },
          "remainingCapacity": {
            "type": "integer",
            "minimum": 0,
            "nullable": true,
            "description": "场次当前剩余总容量；未开售时不返回"
          },
          "ticketTypes": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/SellableTicketType"
            }
          }
        }
      },
      "SellableTicketType": {
        "type": "object",
        "required": [
          "sessionTicketTypeId",
          "ticketTypeName",
          "salePrice"
        ],
        "properties": {
          "sessionTicketTypeId": {
            "type": "integer",
            "format": "int64",
            "description": "场次票种配置 ID，创建订单时提交"
          },
          "ticketTypeName": {
            "type": "string",
            "description": "票种名称"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "票种说明"
          },
          "audienceRule": {
            "type": "string",
            "nullable": true,
            "description": "适用人群规则"
          },
          "salePrice": {
            "type": "number",
            "format": "double",
            "minimum": 0,
            "description": "该场次实际售价"
          },
          "remainingQuantity": {
            "type": "integer",
            "minimum": 0,
            "nullable": true,
            "description": "该场次票种当前剩余数量；未开售时不返回"
          }
        }
      },
      "VenuePage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VenueCard"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "size": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "VenuePageApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/VenuePage"
              }
            }
          }
        ]
      },
      "VenueDetailApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/VenueDetail"
              }
            }
          }
        ]
      },
      "NearbyVenueListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "required": ["data"],
            "properties": {
              "data": {
                "type": "array",
                "items": {"$ref": "#/components/schemas/NearbyVenue"}
              }
            }
          }
        ]
      },
      "SessionListApiResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SellableSession"
                }
              }
            }
          }
        ]
      },
      "ApiResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        }
      },
      "ErrorApiResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "INVALID_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "请求参数不合法"
          }
        }
      }
    },
    "responses": {
      "InvalidRequest": {
        "description": "分页、坐标、搜索半径、地点名称、景点 ID 或参观日期参数不合法",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": "INVALID_REQUEST",
              "message": "请求参数不合法"
            }
          }
        }
      },
      "VenueNotFound": {
        "description": "景点不存在或已经停用",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorApiResponse"
            },
            "example": {
              "code": "NOT_FOUND",
              "message": "景点不存在"
            }
          }
        }
      }
    }
  }
}